Add files via upload
My partner and I made this project for Program Construction at the University of West Georgia in Fall of 2019. To be honest, he did the lion's share of the work on this. I mostly worked on formatting and problem solving if he ran into any issues. Keep an eye on him (mpavich) if you're looking for a hard worker who takes initiative.